fix!(hitbtc): hitbtc3 renamed to hitbtc
the hitbtc class now uses the v3 api, the hitbtc3 class remains as an alias
- copied code in hitbtc3.ts to hitbtc.ts
    - changed all occurences of the text "hitbtc3" to "hitbtc"
- changed hitbtc3 class to an alias for hitbtc

fixes: #18309
